+++ src/stores/konva/stage.ts
@@ -5,7 +5,7 @@ const useStoreStage = defineStore({
   state: () => ({
     configKonva: {
       width: window.innerWidth,
-      height: window.innerHeight / 2,
+      height: window.innerHeight / 1.5,
       scale: {
         x: 1,
         y: 1,
@@ -18,9 +18,9 @@ const useStoreStage = defineStore({
       const containerWidth = container.offsetWidth
       const scale = containerWidth / this.configKonva.width
       this.configKonva.width *= scale
-      this.configKonva.height *= scale
+      // this.configKonva.height *= scale
       this.configKonva.scale.x = scale
-      this.configKonva.scale.y = scale
+      // this.configKonva.scale.y = scale
     },
   },
 })
