+++ src/views/HomePage.vue
@@ -1,7 +1,9 @@
+<script setup lang="ts"></script>
+
 <template lang="pug">
 div(class="my-8 grid gap-4 xl:grid-cols-3 md:grid-cols-2")
   div(v-for="n of 12" :key="n" class="picture m-auto")
-    a(href="")
+    router-link(to =`/create`)
       div(class="bg-gray-200 rounded-lg" style="width: 320px; height: 180px")
     div(class="flex mt-2")
       div(class="mini-avatar bg-gray-200")
